Reduce the time spend on updating the status bar.
An awful lot of time is spend on updating the status bar when
downloading messages. Especially when a folder is almost completely
available locally, with only a few messages that need to be fetched.
In that case the update of the status bar will dominate the downloading.
Updating only every 128th message reduced the download time in my
set-up from 25 minutes to 3 minutes.

Signed-off-by: Attila Kinali <attila@kinali.ch>